// src/enums.ts
export enum ProductType {
  Collection = 'Коллекционная',
  Moment = 'Моментальная',
  Digital = 'Цифровая',
  Film = 'Плёночная',
}

export enum ProductCategory {
  Video = 'Видеокамера',
  Photo = 'Фотоаппарат',
}

export enum ProductLevel {
  Zero = 'Нулевой',
  Amateur = 'Любительский',
  Professional = 'Профессиональный',
}